requirements.txt
asgiref==3.5.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.27
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
idna==3.3
imageio==2.14.1
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msrest==0.6.21
numpy==1.22.1
oauthlib==3.1.1
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.2
environ==1.0